third_party/blink/web_tests/fast/hidpi/2x/image-srcset-intrinsic-size-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById("foo").clientWidth is 200
PASS document.getElementById("foo2").clientWidth is 400
PASS document.getElementById("foo3").clientWidth is 100
PASS document.getElementById("foo4").clientWidth is 200
PASS document.getElementById("foo5").clientWidth is 200
This test passes if the images are all displayed with appropriate dimensions.
